python/requirements.txt
requests
python-dotenv
feedparser